Add method to get jenkins jobs in a folder
To be used for gathering branch jobs for deletion. These need to be
placed in a separate folder to enable a comparison with the branches
found in the git repo.
A pitfall here is to ensure that jobs which are not branches are not
deleted, so jenkins will need to have a folder exclusively for branch
jobs for a given project.